refactor: import 변경 및 사용하지 않는 파라미터 제거
