+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/member/dto/response/MemberResponseDto.java
@@ -2,7 +2,6 @@ package com.mobble.mobbleserver.infrastructure.web.member.dto.response;
 
 import com.mobble.mobbleserver.domain.member.Gender;
 import com.mobble.mobbleserver.domain.member.Member;
-import com.mobble.mobbleserver.infrastructure.web.ground.dto.response.GroundResponseDto;
 
 public record MemberResponseDto(
         Long memberId,
@@ -10,7 +9,6 @@ public record MemberResponseDto(
         Gender gender,
         String email,
         String phone,
-        GroundResponseDto ground,
         String profileImage
 ) {
 
@@ -21,8 +19,6 @@ public record MemberResponseDto(
                 member.getGender(),
                 member.getEmail(),
                 member.getPhone(),
-//                GroundResponseDto.toDto(member.getGround()),
-                null,
                 member.getProfileImage()
         );
     }
